back/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
httpx==0.26.0
python-dotenv==1.0.1
pymongo==4.6.1
qdrant-client==1.8.0
sentence-transformers==3.0.0
PyPDF2==3.0.1
python-docx==1.1.0
numpy>=1.24.0 
google-generativeai==0.8.5